+++ internal/services/migration/azapi_data_plane_resource_migration_v0_to_v1.go
@@ -61,6 +61,7 @@ func AzapiDataPlaneResourceMigrationV0ToV1(ctx context.Context) resource.StateUp
 			Blocks: map[string]schema.Block{
 				"timeouts": timeouts.Block(ctx, timeouts.Opts{
 					Create: true,
+					Update: true,
 					Read:   true,
 					Delete: true,
 				}),