+++ internal/services/migration/azapi_resource_action_migration_v0_to_v1.go
@@ -60,6 +60,7 @@ func AzapiResourceActionMigrationV0ToV1(ctx context.Context) resource.StateUpgra
 			Blocks: map[string]schema.Block{
 				"timeouts": timeouts.Block(ctx, timeouts.Opts{
 					Create: true,
+					Update: true,
 					Read:   true,
 					Delete: true,
 				}),